backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-filter>=23.0
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# JWT Authentication
djangorestframework-simplejwt>=5.3.0

# Rate Limiting
django-ratelimit>=4.1.0

# HTTP Client for IPFS
httpx>=0.27.0

